# Elide Generator::State allocation until a to_json method has to be called