… lib/json/ext/generator/state.rb
… lib/json/pure/generator.rb
… test/json/json_generator_test.rb